Combine storage/v2/postgres and storage/postgres packages and break up giant postgres.go
