+++ components/authz-service/storage/postgres/postgres.go
@@ -4,14 +4,18 @@ import (
 	"context"
 	"database/sql"
 	"strings"
+	"sync"
 
 	"github.com/lib/pq"
 	"github.com/pkg/errors"
 
+	"github.com/chef/automate/components/authz-service/engine"
 	"github.com/chef/automate/components/authz-service/storage"
 	"github.com/chef/automate/components/authz-service/storage/postgres/datamigration"
 	"github.com/chef/automate/components/authz-service/storage/postgres/migration"
 	"github.com/chef/automate/lib/db"
+	"github.com/chef/automate/lib/grpc/auth_context"
+	"github.com/chef/automate/lib/logger"
 )
 
 // New returns a new sql.DB connector with the automatic default migrations applied.
@@ -100,4 +104,127 @@ func parsePQError(e *pq.Error) error {
 	}
 
 	return storage.ErrDatabase
+}
+
+type pg struct {
+	db           *sql.DB
+	engine       engine.Engine
+	logger       logger.Logger
+	dataMigConf  datamigration.Config
+	conninfo     string
+	projectLimit int
+}
+
+var singletonInstance *pg
+var once sync.Once
+
+type Querier interface {
+	ExecContext(context.Context, string, ...interface{}) (sql.Result, error)
+	QueryContext(context.Context, string, ...interface{}) (*sql.Rows, error)
+	QueryRowContext(context.Context, string, ...interface{}) *sql.Row
+}
+
+// GetInstance returns the singleton instance. Will be nil if not yet initialized.
+func GetInstance() *pg {
+	return singletonInstance
+}
+
+// Initialize instantiates the singleton postgres storage backend.
+// Will only initialize once. Will simply return nil if already initialized.
+func Initialize(ctx context.Context, e engine.Engine, l logger.Logger, migConf migration.Config,
+	dataMigConf datamigration.Config, projectLimit int) error {
+
+	var err error
+	once.Do(func() {
+		l.Infof("applying database migrations from %s", migConf.Path)
+		var db *sql.DB
+		db, err = New(ctx, migConf, dataMigConf)
+		singletonInstance = &pg{
+			db:           db,
+			engine:       e,
+			logger:       l,
+			dataMigConf:  dataMigConf,
+			conninfo:     migConf.PGURL.String(),
+			projectLimit: projectLimit}
+	})
+	return err
+}
+
+func (p *pg) notifyPolicyChange(ctx context.Context, q Querier) error {
+	// We keep track of an id with each change. This lets us be smart about only updating
+	// the OPA rules when it might change.
+	_, err := q.ExecContext(ctx, "SELECT notify_policy_change()")
+	return err
+}
+
+func (p *pg) Reset(ctx context.Context) error {
+	if _, err := p.db.ExecContext(ctx,
+		`TRUNCATE TABLE iam_policies, iam_members, iam_roles, iam_projects CASCADE;`); err != nil {
+		return errors.Wrap(err, "truncate database")
+	}
+
+	if err := p.dataMigConf.Reset(); err != nil {
+		return errors.Wrap(err, "reset v2 data migrations")
+	}
+
+	return nil
+}
+
+func (p *pg) Close() error {
+	err := errors.Wrap(p.db.Close(), "close database connection")
+	// reset the singleton
+	once = *new(sync.Once)
+	singletonInstance = nil
+	return err
+}
+
+func (p *pg) singleRowResultOrNotFoundErr(result sql.Result) error {
+	count, err := result.RowsAffected()
+	if err != nil {
+		return p.processError(err)
+	}
+	if count == 0 {
+		return storage.ErrNotFound
+	}
+	if count > 1 {
+		return storage.ErrDatabase
+	}
+	return nil
+}
+
+func (p *pg) getMapOfRuleAssociations(ctx context.Context, q Querier, id string, projectID string) (map[string]bool, error) {
+	assocRow := q.QueryRowContext(ctx, "SELECT query_rule_table_associations($1, $2)", id, projectID)
+	var associations []string
+	if err := assocRow.Scan(pq.Array(&associations)); err != nil {
+		return nil, err
+	}
+
+	set := make(map[string]bool, len(associations))
+	for _, s := range associations {
+		set[s] = true
+	}
+	return set, nil
+}
+
+func (p *pg) processError(err error) error {
+	p.logger.Debugf("err: %v", err)
+	err = ProcessError(err)
+	if err == storage.ErrDatabase {
+		p.logger.Warnf("unknown error type from database: %v", err)
+	}
+	return err
+}
+
+// projectsListFromContext returns the project list from the context.
+// In the case that the project list was ["*"], we return an empty list,
+// since we do not wish to filter on projects.
+func projectsListFromContext(ctx context.Context) ([]string, error) {
+	projectsFilter, err := auth_context.ProjectsFromIncomingContext(ctx)
+	if err != nil {
+		return nil, err
+	}
+	if auth_context.AllProjectsRequested(projectsFilter) {
+		projectsFilter = []string{}
+	}
+	return projectsFilter, nil
 }
